requirements.txt
# RAG Chatbot Dependencies

# Core framework
streamlit>=1.28.0

# LangChain components
langchain>=0.1.0
langchain-community>=0.0.20

# Vector store and embeddings
chromadb>=0.4.0

# HTTP requests
requests>=2.28.0

# PDF processing
pypdf>=3.0.0

# Environment variables (optional)
python-dotenv>=1.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Type hints
typing-extensions>=4.5.0